Lazily attach the svg filter to the page
the content script loads the svg but doesn't attach it until the extension window tells it to. When you select a colour profile for the first time the extension will tell the content script to load.

This prevents the extension from interfering with pages until actually used. This fixes the known issue with bitbucket in #3 and prevents further issues.